fix(dictation): show readiness setup on settings page
